Should be fixed
Fuel cell had several issues: steam to carbon equation was wrong (and
wrong in paper), # of cells determines average densit and thus current,
so it was causing problems, ASR should be ~.25 not 2, Energy balance was
never quite right, I have simplified it to Q gen = Q reform.

Turbine was bad because when we were using total enthalpy the negatives
were throwing off how the isentropic efficiency equation fins Hout. It
wasn't a problem in compressor because hf of O2 and N2 is 0.

Generally shortened main function to avoid duplication

Second compressor was being fed O2 at 1023K, need to intercool